fix various unsync data bug due to switching library slot and tag and cache update

+++ src/Tagcade/DomainManager/LibraryExpressionManager.php
@@ -50,6 +50,8 @@ class LibraryExpressionManager implements LibraryExpressionManagerInterface
      */
     public function delete(ModelInterface $libraryExpression)
     {
+        if(!$libraryExpression instanceof LibraryExpressionInterface) throw new InvalidArgumentException('expect LibraryExpressionInterface object');
+
         $this->em->remove($libraryExpression);
         $this->em->flush();
     }
